+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v6.3.0 (unreleased)
+This release extends the planned support of the collections to OneView REST API version 3200 (OneView v6.3).
+
+### Modules supported in this release
+- oneview_appliance_proxy_configuration
+- oneview_appliance_proxy_configuration_facts
+
 # Ansible Collections for HPE OneView Change Log
 
 ## v6.2.0